requirements.txt
flask==2.3.3
pandas==2.3.1
plotly==5.15.0
jinja2==3.1.2
werkzeug==2.3.7
python-dotenv==1.0.0
requests>=2.31.0